+++ waku/v2/rpc/store.go
@@ -27,8 +27,8 @@ type StorePagingOptions struct {
 type StoreMessagesArgs struct {
 	Topic          string             `json:"pubsubTopic,omitempty"`
 	ContentFilters []string           `json:"contentFilters,omitempty"`
-	StartTime      float64            `json:"startTime,omitempty"`
-	EndTime        float64            `json:"endTime,omitempty"`
+	StartTime      int64              `json:"startTime,omitempty"`
+	EndTime        int64              `json:"endTime,omitempty"`
 	PagingOptions  StorePagingOptions `json:"pagingOptions,omitempty"`
 }
